Function communication actually returns objects.
subscribeStatistics now actually sends the data to the ui (although it does not show it in the ui yet, but it shows receiving it in console).

getStaticData now also usable from ui (await window.electron.getStaticData()) This will return the getStaticData data. (once again only in console, as we are not showing it in the ui yet.